refactor: 유효성 검증 함수 역할에 따라 분리
